requirements.txt
flask
openai
python-dotenv
google-cloud-dialogflow
